Moved create itinerary form to separate page
