+++ resources/pages/CreateItinerary/DeleteItineraryModal.tsx
