+++ resources/pages/Home/Home.tsx
@@ -1,15 +1,7 @@
-import { CreateItineraryModal } from './CreateItineraryModal';
 import { WelcomeCard } from './WelcomeCard';
 
-export interface HomePageNavigationState {
-  createItinerary: boolean | undefined;
-}
-
 export const Home = (): JSX.Element => (
-  <>
-    <div className="flex flex-1 flex-col p-2">
-      <WelcomeCard />
-    </div>
-    <CreateItineraryModal />
-  </>
+  <div className="flex flex-1 flex-col p-2">
+    <WelcomeCard />
+  </div>
 );
